separate business logic from Filter component
